better documentation
