+++ docs/examples/docs-demo-six-datepicker-19.vue
@@ -1,11 +1,13 @@
 <template>
 <div>
 
-        <six-datepicker size="small" clearable></six-datepicker>
-        <br>
-        <six-datepicker size="medium" clearable></six-datepicker>
-        <br>
-        <six-datepicker size="large" clearable></six-datepicker>
+        <six-datepicker custom-icon>
+          <img             style="height: 24px"
+            slot="custom-icon"
+            src="https://upload.wikimedia.org/wikipedia/commons/thumb/5/59/Sketch_Logo.svg/1133px-Sketch_Logo.svg.png"
+            alt="customicon"
+          >
+        </six-datepicker>
         
       
 </div>
